#235: validate reaction
