+++ src/domain/reaction/create/InvalidReaction.ts
@@ -1,0 +1,7 @@
+
+import { ValidationError } from '^/integrations/runtime/module';
+
+export default class InvalidReaction extends ValidationError
+{
+
+}
