+++ src/domain/reaction/create/types.ts
@@ -1,0 +1,6 @@
+
+import { DataModel } from '../types';
+
+type ValidationModel = Pick<DataModel, 'postId' | 'comicId' | 'commentId'>;
+
+export type { ValidationModel };
