+++ test/domain/reaction/createComic.spec.ts
@@ -2,13 +2,10 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 
 import { RECORD_TYPE as COMIC_RECORD_TYPE } from '^/domain/comic/definitions';
-import InvalidComment from '^/domain/comment/create/InvalidComment';
-import { RECORD_TYPE as COMMENT_RECORD_TYPE } from '^/domain/comment/definitions';
 import { RECORD_TYPE as IMAGE_RECORD_TYPE } from '^/domain/image/definitions';
 import PostNotFound from '^/domain/post/PostNotFound';
 import { RECORD_TYPE as POST_RECORD_TYPE } from '^/domain/post/definitions';
 import createComicReaction from '^/domain/reaction/createComic/feature';
-import createCommentReaction from '^/domain/reaction/createComment/feature';
 import { RECORD_TYPE as REACTION_RECORD_TYPE } from '^/domain/reaction/definitions';
 
 import database from '^/integrations/database/module';
@@ -26,28 +23,6 @@ beforeEach(async () =>
 
 describe('domain/reaction/create', () =>
 {
-    it('should create a comment reaction', async () =>
-    {
-        const reactionId = await createCommentReaction(REQUESTERS.OWNER, VALUES.IDS.POST_EXISTING, VALUES.MESSAGES.VALID_COMMENT);
-
-        const reaction = await database.readRecord(REACTION_RECORD_TYPE, reactionId);
-        expect(reaction?.creatorId).toBe(REQUESTERS.OWNER.id);
-        expect(reaction?.postId).toBe(VALUES.IDS.POST_EXISTING);
-        expect(reaction?.comicId).toBeUndefined();
-        expect(reaction?.ratingCount).toBe(0);
-        expect(reaction?.createdAt).toBeDefined();
-
-        const post = await database.readRecord(POST_RECORD_TYPE, reaction.postId as string);
-        expect(post?.creatorId).toBe(REQUESTERS.OWNER.id);
-        expect(post?.comicId).toBeDefined();
-        expect(post?.createdAt).toBeDefined();
-        expect(post?.ratingCount).toBe(0);
-        expect(post?.reactionCount).toBe(1);
-
-        const comment = await database.readRecord(COMMENT_RECORD_TYPE, reaction.commentId as string);
-        expect(comment?.message).toBe(VALUES.MESSAGES.VALID_COMMENT);
-    });
-
     it('should create a comic reaction', async () =>
     {
         const reactionId = await createComicReaction(REQUESTERS.OWNER, VALUES.IDS.POST_EXISTING, VALUES.DATA_URLS.COMIC);
@@ -74,38 +49,6 @@ describe('domain/reaction/create', () =>
 
         const file = await fileStore.readFile(image.storageKey as string);
         expect(file).toBeDefined();
-    });
-
-    it('should rollback created data at non-existing post comment reaction', async () =>
-    {
-        // This should fail at the last action when incrementing post's reaction count
-        const promise = createCommentReaction(REQUESTERS.OWNER, VALUES.IDS.POST_NOT_EXISTING, VALUES.MESSAGES.VALID_COMMENT);
-        await expect(promise).rejects.toThrow(PostNotFound);
-
-        const reactions = await database.searchRecords(REACTION_RECORD_TYPE, {});
-        expect(reactions).toHaveLength(5);
-
-        const posts = await database.searchRecords(POST_RECORD_TYPE, {});
-        expect(posts).toHaveLength(1);
-
-        const comments = await database.searchRecords(COMMENT_RECORD_TYPE, {});
-        expect(comments).toHaveLength(1);
-    });
-
-    it('should rollback created data at invalid comment reaction', async () =>
-    {
-        // This should fail at the first action when creating the comment
-        const promise = createCommentReaction(REQUESTERS.OWNER, VALUES.IDS.POST_EXISTING, VALUES.MESSAGES.INVALID_COMMENT);
-        await expect(promise).rejects.toThrow(InvalidComment);
-
-        const reactions = await database.searchRecords(REACTION_RECORD_TYPE, {});
-        expect(reactions).toHaveLength(5);
-
-        const posts = await database.searchRecords(POST_RECORD_TYPE, {});
-        expect(posts).toHaveLength(1);
-
-        const comments = await database.searchRecords(COMMENT_RECORD_TYPE, {});
-        expect(comments).toHaveLength(1);
     });
 
     it('should rollback created data at failed comic reaction', async () =>
